core: Split up serialization into core and utils
