+++ src/core/cuda_interface.cpp
@@ -27,9 +27,10 @@
 #include "energy.hpp"
 #include "grid.hpp"
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
+#include "serialization/CUDA_particle_data.hpp"
+
 #include "utils/mpi/gather_buffer.hpp"
 #include "utils/mpi/scatter_buffer.hpp"
-#include "utils/serialization/CUDA_particle_data.hpp"
 
 /// MPI tag for cuda particle gathering
 #define REQ_CUDAGETPARTS 0xcc01
