+++ src/core/nonbonded_interactions/nonbonded_interaction_data.cpp
@@ -64,9 +64,8 @@
 #include "pressure.hpp"
 #include "rattle.hpp"
 #include "reaction_field.hpp"
+#include "serialization/IA_parameters.hpp"
 #include "thermostat.hpp"
-#include "utils.hpp"
-#include "utils/serialization/IA_parameters.hpp"
 
 #include <boost/archive/binary_iarchive.hpp>
 #include <boost/archive/binary_oarchive.hpp>
